shifting-appropriate-methods-to-auth-lib
