+++ lib/msf/core/exploit/remote/ldap/server.rb
@@ -9,6 +9,7 @@ module Msf
   module Exploit::Remote::LDAP
     module Server
       include Exploit::Remote::SocketServer
+      include Rex::Proto::LDAP::Server::Auth
 
       #
       # Initializes an exploit module that serves LDAP requests
